SCUMM HE: Moonbase now playable via local network.
This requires the ENet library to be compiled and enabled to access
these features.

This only works over local network for now.  Internet-wide support
(with hole-punching to bypass NAT) will be implemented later down the
line.

There are some bugs currently that needs squashing and some cleanup
are warranted, but I'll get to them.